refactor(core): move action id to constants
